Switch from wchar_t to char
JSON is a language defined by unicode code points. However, all the
"control characters" of JSON are valid ASCII: punctuation, ASCII
whitespace, and ASCII digits. In the Linux/Unix world, the vast majority
of all text is encoded UTF-8. While the handling of text encodings in
this library is a technically correct one ("I don't know anything about
your application's text encodings, so please give me decoded 4-byte
characters"), practically it is incredibly painful.

For one, the vast majority of cases can be served by an application
which is agnostic to encodings, and "pretends" that text is ASCII (but
it is really UTF-8). At the point that an application needs to
understand text encodings, it can always decode the (much smaller) text
of interest. Inserting nosj into an application which already handles
text (and, indeed, Unicode) reasonably well is a pain, because one must
convert UTF-8 to wchar_t and back again.

For another, a major selling point of NOSJ is its low footprint. It is a
zero-allocation library. It is not the fastest, but it is correct and
capable of parsing the structure of the text so that an application can
extract the minimum data it needs, with minimal memory allocations. It
seems conceivable that this approach would be incredibly useful on an
embedded system, or for very simple JSON. Those applications are the
exact set of applications which would be very frustrated by the extra
conversions (and memory) required to decode data which likely was mostly
ASCII anyway.

SO. This commit changes the entire NOSJ library to accept char strings.
This is a mechanical commit, and it doesn't completely make a full
migration. There are some compile warnings in src/string.c due to
comparing char to integers out of range (this is the unicode escape
stuff) and then the surrogate pairs test fails. This is all to be
expected, as those portions of the code involve actual logic reasoning
about Unicode. I will address these in upcoming commits.